Reorganizacion de la estructura del codigo
